Refactor code and create new components
