+++ pruebas/01-reading-list/RANDRESS23/src/components/BookShelf/styles/BookShelf.module.css
@@ -16,397 +16,7 @@
   /* height: 100%; */
 }
 
-.ShelfSectionPendingBooks {
-  width: 100%;
-  display: flex;
-  /* background-color: blue; */
-}
-
-.ShelfPendingBooks {
-  position: relative;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  width: 40%;
-}
-
-.TitlesContainer {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  gap: 20px;
-}
-
-.TitleShelf {
-  font-size: 50px;
-  font-family: 'Libre Caslon Text', serif;
-  /* color: #daaa63; */
-  color: #2c1810;
-}
-
-.SubTitleShelf {
-  font-size: 16px;
-  font-family: 'Montserrat', sans-serif;
-  /* color: #daaa63; */
-  color: #2c1810;
-}
-
-.BookSign {
-  position: absolute;
-  z-index: 10;
-  bottom: -10%;
-}
-
-.Sign {
-  position: relative;
-  z-index: 10;
-  background-color: #daaa63;
-  color: #2c1810;
-  font-weight: bold;
-  padding: 10px 20px;
-  border-radius: 3px;
-  font-family: 'Montserrat', sans-serif;
-  box-shadow: rgba(50, 50, 93, 0.25) 0px 50px 100px -20px, rgba(0, 0, 0, 0.3) 0px 30px 60px -30px, rgba(10, 37, 64, 0.35) 0px -2px 6px 0px inset;
-}
-
-.CarouselBooks {
-  width: 57%;
-  position: relative;
-  z-index: 10;
-  display: flex;
-  cursor: grab;
-  overflow: hidden;
-  transform: translateY(10%);
-}
-
-.BooksContainer {
-  display: flex;
-  padding-right: 30px;
-  gap: 30px;
-}
-
-.BookReading {
-  position: relative;
-  display: flex;
-  /* width: 120px;
-  height: 180px; small size */
-  width: 180px;
-  height: 240px;
-}
-
-.BookCover {
-  position: relative;
-  z-index: 10;
-  width: 96%;
-  height: 100%;
-  border-radius: 8px;
-  border-top-right-radius: 5px;
-  border-bottom-right-radius: 5px;
-  pointer-events: none;
-}
-
-.BookColumn {
-  position: absolute;
-  z-index: 10;
-  top: 0;
-  left: 0;
-  width: 7%;
-  height: 100%;
-  border-radius: 8px;
-  border-top-right-radius: 0;
-  border-bottom-right-radius: 0;
-  box-shadow: inset rgba(0, 0, 0, 0.56) 0px 3px 5px 2px;
-}
-
-.BorderSheets {
-  position: relative;
-  z-index: 10;
-  height: 100%;
-  width: 4%;
-  background-color: #aea190;
-  border: 1px solid #aea190;
-  clip-path: polygon(0 2%, 0 2%, 100% 10%, 100% 90%, 0 98%, 0 98%, 0% 70%, 0% 30%);
-}
-
-.Sheets {
-  width: 100%;
-  height: 100%;
-  background-color: #f3f1ee;
-  clip-path: polygon(0 4%, 0 4%, 100% 10%, 100% 90%, 0 96%, 0 96%, 0 65%, 0% 30%);
-}
-
-.BookShadow {
-  top: 0;
-  left: -5%;
-  width: 165%;
-  height: 145%;
-  position: absolute;
-  z-index: 1;
-  clip-path: polygon(25% 0, 58% 0, 79% 40%, 100% 76%, 100% 100%, 30% 100%, 16% 70%, 18% 31%);
-  box-shadow: rgba(0, 0, 0, 0.56) 0px 22px 140px 4px inset;
-}
-
-.ShelfTable {
-  position: relative;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  width: 100%;
-  height: 60px;
-}
-
-.Table {
-  position: relative;
-  width: 110%;
-  height: 75%;
-  background-color: #e9e0d2;
-  clip-path: polygon(5% 0, 95% 0, 100% 100%, 0% 100%);
-  border: 1px solid #e9e0d2;
-  box-shadow: inset rgba(100, 100, 111, 0.2) 0px 7px 29px 0px;
-}
-
-.BorderTable {
-  position: relative;
-  width: 110%;
-  height: 25%;
-  background-color: #f6f2eb;
-  border-radius: 3px;
-  border: 1px solid #e9e0d2;
-}
-
-.ShadowTable {
-  top: 100%;
-  left: 0;
-  position: absolute;
-  width: 105%;
-  height: 120%;
-  clip-path: polygon(0 0, 98% 0, 100% 100%, 3% 100%);
-  box-shadow: rgba(0, 0, 0, 0.45) 0px 25px 25px -20px inset;
-}
-
 .ShelfDown {
   width: 100%;
   margin-top: 70px;
-}
-
-.TitlesContainerCount { 
-  padding: 5px 30px;
-  margin-bottom: 30px;
-  display: flex;
-  flex-direction: column;
-  gap: 15px;
-}
-
-.TitleContent {
-  display: flex;
-  align-items: center;
-  gap: 15px;
-}
-
-.TitleBooksAvailable {
-  font-size: 30px;
-  font-family: 'Libre Caslon Text', serif;
-  color: #2c1810;
-}
-
-.TitleGenreBooksAvailable {
-  font-size: 20px;
-  font-family: 'Libre Caslon Text', serif;
-  color: #2c1810;
-}
-
-.SignAvailable {
-  position: relative;
-  z-index: 10;
-  background-color: #daaa63;
-  color: #2c1810;
-  font-weight: bold;
-  font-size: 20px;
-  font-family: 'Montserrat', sans-serif;
-  padding: 10px 20px;
-  border-radius: 3px;
-  box-shadow: rgba(50, 50, 93, 0.25) 0px 50px 100px -20px, rgba(0, 0, 0, 0.3) 0px 30px 60px -30px, rgba(10, 37, 64, 0.35) 0px -2px 6px 0px inset;
-}
-
-.SignGenreAvailable {
-  position: relative;
-  z-index: 10;
-  background-color: #daaa63;
-  color: #2c1810;
-  font-weight: bold;
-  font-size: 16px;
-  font-family: 'Montserrat', sans-serif;
-  padding: 7px 15px;
-  border-radius: 3px;
-  box-shadow: rgba(50, 50, 93, 0.25) 0px 50px 100px -20px, rgba(0, 0, 0, 0.3) 0px 30px 60px -30px, rgba(10, 37, 64, 0.35) 0px -2px 6px 0px inset;
-}
-
-.FiltersSection {
-  padding: 10px 30px;
-  display: flex;
-  /* background-color: blue; */
-  gap: 40px;
-}
-
-.Filter {
-  position: relative;
-  display: flex;
-  /* justify-content: space-between; */
-  flex-direction: column;
-  gap: 10px;
-  /* background-color: red; */
-}
-
-.IconContainer {
-  position: absolute;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  height: 45px;
-  width: 40px;
-}
-
-.InputSearch {
-  background-color: #fff;
-  color: #2c1810;
-  border: none;
-  padding: 12px 40px;
-  border-radius: 30px;
-  font-family: 'Montserrat', sans-serif;
-  font-size: 15px;
-  box-shadow: rgba(0, 0, 0, 0.1) 0px 10px 50px;
-}
-
-.InputSearch:focus {
-  outline: none;
-}
-
-.FilterLabel {
-  font-family: 'Montserrat', sans-serif;
-  font-size: 15px;
-  color: #2c1810;
-}
-
-.FilterSelectContainer {
-  position: relative;
-  display: flex;
-  width: 250px;
-  height: 40px;
-  border-radius: 30px;
-  overflow: hidden;
-  box-shadow: rgba(0, 0, 0, 0.1) 0px 10px 50px;
-}
-
-.FilterSelectContainer::after {
-  content: '\25BC';
-  position: absolute;
-  top: 0;
-  right: 0;
-  padding: 1em;
-  background-color: #daaa63;
-  color: #2c1810;
-  transition: .25s all ease;
-  pointer-events: none;
-}
-
-.FilterSelectContainer:hover::after {
-  color: #5c3323;
-}
-
-.FilterSelect {
-  color: #2c1810;
-  font-size: 15px;
-  appearance: none;
-  border: none;
-  flex: 1;
-  padding: 0 20px;
-  color: #2c1810;
-  background-color: #fff;
-  cursor: pointer;
-}
-
-.FilterSelect::-ms-expand {
-  display: none;
-}
-
-.FilterSelect:focus {
-  outline: none;
-}
-
-.FilterOption {
-  border-radius: 30px;
-}
-
-.InputRange {
-  -webkit-appearance: none;
-  appearance: none; 
-  width: 200px;
-  cursor: pointer;
-  outline: none;
-  overflow: hidden;
-  border-radius: 16px;
-}
-
-.InputRange::-webkit-slider-runnable-track {
-  height: 15px;
-  background: #daaa63;
-  border-radius: 16px;
-}
-
-.InputRange::-webkit-slider-thumb {
-  -webkit-appearance: none;
-  appearance: none; 
-  height: 15px;
-  width: 15px;
-  background-color: #fff;
-  border-radius: 50%;
-  border: 2px solid #2c1810;
-  box-shadow: -407px 0 0 400px #2c1810;
-}
-
-.InputRange:active {
-	cursor: grabbing;
-}
-
-.InputRange:disabled {
-	filter: grayscale(1);
-	opacity: 0.3;
-	cursor: not-allowed;
-}
-
-.BooksAvailableContainer {
-  padding: 30px;
-}
-
-.BookAvailableReading {
-  position: relative;
-  display: flex;
-  width: 130px;
-  height: 190px;
-}
-
-.BookAvailableCover {
-  position: relative;
-  z-index: 10;
-  width: 96%;
-  height: 100%;
-  border-radius: 8px;
-  border-top-right-radius: 5px;
-  border-bottom-right-radius: 5px;
-  cursor: pointer;
-}
-
-.BookAvailableShadow {
-  top: 0;
-  left: -5%;
-  width: 165%;
-  height: 115%;
-  position: absolute;
-  z-index: 1;
-  clip-path: polygon(25% 0, 58% 0, 79% 40%, 100% 76%, 100% 100%, 30% 100%, 16% 70%, 18% 31%);
-  background: rgb(0,0,0);
-  background: linear-gradient(325deg, rgba(0,0,0,0) 20%, rgba(0,0,0,0.12368697478991597) 43%, rgba(0,0,0,0.23853291316526615) 61%, rgba(0,0,0,0.3897934173669467) 77%, rgba(0,0,0,0.7091211484593838) 100%);
 }
