+++ pruebas/01-reading-list/RANDRESS23/src/components/BookSign/BookSign.tsx
@@ -1,0 +1,9 @@
+import styles from './styles/BookSign.module.css'
+
+export const BookSign: React.FC = () => {
+  return (
+    <div className={styles.BookSign}>
+      <div className={styles.Sign}>Books to read: 10</div>
+    </div>
+  )
+}
