+++ pruebas/01-reading-list/RANDRESS23/src/components/BookSign/styles/BookSign.module.css
@@ -1,0 +1,17 @@
+.BookSign {
+  position: absolute;
+  z-index: 10;
+  bottom: -10%;
+}
+
+.Sign {
+  position: relative;
+  z-index: 10;
+  background-color: #daaa63;
+  color: #2c1810;
+  font-weight: bold;
+  padding: 10px 20px;
+  border-radius: 3px;
+  font-family: 'Montserrat', sans-serif;
+  box-shadow: rgba(50, 50, 93, 0.25) 0px 50px 100px -20px, rgba(0, 0, 0, 0.3) 0px 30px 60px -30px, rgba(10, 37, 64, 0.35) 0px -2px 6px 0px inset;
+}
