+++ pruebas/01-reading-list/RANDRESS23/src/components/BooksAvailable/BooksAvailable.tsx
@@ -1,0 +1,10 @@
+import { Book } from '.'
+import styles from './styles/BooksAvailable.module.css'
+
+export const BooksAvailable: React.FC = () => {
+  return (
+    <div className={styles.BooksAvailable}>
+      <Book />
+    </div>
+  )
+}
