+++ pruebas/01-reading-list/RANDRESS23/src/components/BooksAvailable/components/Book/index.ts
@@ -1,0 +1,1 @@
+export * from './Book'
