+++ pruebas/01-reading-list/RANDRESS23/src/components/BooksAvailable/components/Book/styles/Book.module.css
@@ -1,0 +1,46 @@
+.BookAvailableReading {
+  position: relative;
+  display: flex;
+  width: 130px;
+  height: 190px;
+}
+
+.BookAvailableCover {
+  position: relative;
+  z-index: 10;
+  width: 96%;
+  height: 100%;
+  border-radius: 8px;
+  border-top-right-radius: 5px;
+  border-bottom-right-radius: 5px;
+  cursor: pointer;
+}
+
+.BorderSheets {
+  position: relative;
+  z-index: 10;
+  height: 100%;
+  width: 4%;
+  background-color: #aea190;
+  border: 1px solid #aea190;
+  clip-path: polygon(0 2%, 0 2%, 100% 10%, 100% 90%, 0 98%, 0 98%, 0% 70%, 0% 30%);
+}
+
+.Sheets {
+  width: 100%;
+  height: 100%;
+  background-color: #f3f1ee;
+  clip-path: polygon(0 4%, 0 4%, 100% 10%, 100% 90%, 0 96%, 0 96%, 0 65%, 0% 30%);
+}
+
+.BookAvailableShadow {
+  top: 0;
+  left: -5%;
+  width: 165%;
+  height: 115%;
+  position: absolute;
+  z-index: 1;
+  clip-path: polygon(25% 0, 58% 0, 79% 40%, 100% 76%, 100% 100%, 30% 100%, 16% 70%, 18% 31%);
+  background: rgb(0,0,0);
+  background: linear-gradient(325deg, rgba(0,0,0,0) 20%, rgba(0,0,0,0.12368697478991597) 43%, rgba(0,0,0,0.23853291316526615) 61%, rgba(0,0,0,0.3897934173669467) 77%, rgba(0,0,0,0.7091211484593838) 100%);
+}
