+++ pruebas/01-reading-list/RANDRESS23/src/components/BooksAvailable/index.ts
@@ -1,0 +1,2 @@
+export * from './BooksAvailable'
+export * from './components/Book'
