+++ pruebas/01-reading-list/RANDRESS23/src/components/CarouselBooks/index.ts
@@ -1,0 +1,2 @@
+export * from './CarouselBooks'
+export * from './components/BookReading'
