+++ pruebas/01-reading-list/RANDRESS23/src/components/CarouselBooks/styles/CarouselBooks.module.css
@@ -1,0 +1,15 @@
+.CarouselBooks {
+  width: 57%;
+  position: relative;
+  z-index: 10;
+  display: flex;
+  cursor: grab;
+  overflow: hidden;
+  transform: translateY(10%);
+}
+
+.BooksContainer {
+  display: flex;
+  padding-right: 30px;
+  gap: 30px;
+}
