+++ pruebas/01-reading-list/RANDRESS23/src/components/FiltersSection/FiltersSection.tsx
@@ -1,0 +1,12 @@
+import { InputRange, InputSearch, SelectGenre } from '.'
+import styles from './styles/FiltersSection.module.css'
+
+export const FiltersSection: React.FC = () => {
+  return (
+    <div className={styles.FiltersSection}>
+      <InputSearch />
+      <SelectGenre />
+      <InputRange />
+    </div>
+  )
+}
