+++ pruebas/01-reading-list/RANDRESS23/src/components/FiltersSection/components/InputRange/InputRange.tsx
@@ -1,0 +1,10 @@
+import styles from './styles/InputRange.module.css'
+
+export const InputRange: React.FC = () => {
+  return (
+    <div className={styles.Filter}>
+      <label className={styles.FilterLabel}>Filtrar por paginas</label>
+      <input type="range" className={styles.InputRange} />
+    </div>
+  )
+}
