+++ pruebas/01-reading-list/RANDRESS23/src/components/FiltersSection/components/InputSearch/SearchIcon.tsx
