+++ pruebas/01-reading-list/RANDRESS23/src/components/FiltersSection/components/InputSearch/index.ts
@@ -1,1 +1,2 @@
 export * from './InputSearch'
+export * from './SearchIcon'
