+++ pruebas/01-reading-list/RANDRESS23/src/components/FiltersSection/index.ts
@@ -1,0 +1,4 @@
+export * from './FiltersSection'
+export * from './components/InputRange'
+export * from './components/InputSearch'
+export * from './components/SelectGenre'
