+++ pruebas/01-reading-list/RANDRESS23/src/components/FiltersSection/styles/FiltersSection.module.css
@@ -1,0 +1,5 @@
+.FiltersSection {
+  padding: 10px 30px;
+  display: flex;
+  gap: 40px;
+}
