--- pruebas/01-reading-list/RANDRESS23/src/components/InputSearch/InputSearch.tsx
@@ -1,5 +1,0 @@
-export const InputSearch: React.FC = () => {
-  return (
-    <div>input</div>
-  )
-}
