--- pruebas/01-reading-list/RANDRESS23/src/components/InputSearch/components/index.ts
@@ -1,1 +1,0 @@
-export * from './SearchIcon'
