+++ pruebas/01-reading-list/RANDRESS23/src/components/PendingBooks/styles/PendingBooks.module.css
@@ -1,0 +1,13 @@
+.ShelfSectionPendingBooks {
+  width: 100%;
+  display: flex;
+}
+
+.ShelfPendingBooks {
+  position: relative;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  width: 40%;
+}
