+++ pruebas/01-reading-list/RANDRESS23/src/components/ShelfTable/index.ts
@@ -1,0 +1,1 @@
+export * from './ShelfTable'
