+++ pruebas/01-reading-list/RANDRESS23/src/components/TitlesContainer/TitlesContainer.tsx
@@ -1,0 +1,11 @@
+import { motion } from 'framer-motion'
+import styles from './styles/TitlesContainer.module.css'
+
+export const TitlesContainer: React.FC = () => {
+  return (
+    <div className={styles.TitlesContainer}>
+      <motion.h1 className={styles.TitleShelf}>Reading List</motion.h1>
+      <motion.p className={styles.SubTitleShelf}>Lorem ipsum dolor sit amet.</motion.p>
+    </div>
+  )
+}
