+++ pruebas/01-reading-list/RANDRESS23/src/components/TitlesContainer/styles/TitlesContainer.module.css
@@ -1,0 +1,19 @@
+.TitlesContainer {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  gap: 20px;
+}
+
+.TitleShelf {
+  font-size: 50px;
+  font-family: 'Libre Caslon Text', serif;
+  color: #2c1810;
+}
+
+.SubTitleShelf {
+  font-size: 16px;
+  font-family: 'Montserrat', sans-serif;
+  color: #2c1810;
+}
