+++ pruebas/01-reading-list/RANDRESS23/src/components/TitlesContainerCount/TitlesContainerCount.tsx
@@ -1,0 +1,19 @@
+import { TitleContent } from '.'
+import styles from './styles/TitlesContainerCount.module.css'
+
+export const TitlesContainerCount: React.FC = () => {
+  return (
+    <div className={styles.TitlesContainerCount}>
+      <TitleContent
+        title='Books Available'
+        count={10}
+        isTitleGenre={false}
+      />
+      <TitleContent
+        title='Genre Books'
+        count={5}
+        isTitleGenre={true}
+      />
+    </div>
+  )
+}
