+++ pruebas/01-reading-list/RANDRESS23/src/components/TitlesContainerCount/index.ts
@@ -1,0 +1,2 @@
+export * from './TitlesContainerCount'
+export * from './components/TitleContent'
