+++ pruebas/01-reading-list/RANDRESS23/src/components/TitlesContainerCount/styles/TitlesContainerCount.module.css
@@ -1,0 +1,7 @@
+.TitlesContainerCount { 
+  padding: 5px 30px;
+  margin-bottom: 30px;
+  display: flex;
+  flex-direction: column;
+  gap: 15px;
+}
